backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# AI & Embeddings
anthropic==0.18.1
qdrant-client==1.7.3
tiktoken==0.6.0

# GitHub Integration
PyGithub==2.2.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Code Parsing
tree-sitter==0.23.2
tree-sitter-php==0.23.11